Test cleanup: Move hostdisk utils to an appropriate package
